pyproject.toml
[tool.poetry]
name = "iis-projekt"
version = "0.1.0"
description = ""
authors = ["kristjansuligoj <kristjan.suligoj@aspiria.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
flask-cors = "^4.0.0"
cryptography = "^42.0.7"
pandas = "^2.2.1"
retry-requests = "^2.0.0"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.0"
great-expectations = "^0.18.13"
evidently = "^0.4.19"
litestar="^2.8.2"
mlflow="^2.12.1"
black = "^24.4.2"
jupyter = "^1.0.0"
matplotlib = "^3.9.0"
botocore = "^1.34.111"
dagshub = "^0.3.27"
pymongo= { version = "^3.12", extras=["srv"] }
tf2onnx = "1.16.1"
onnxruntime = "^1.17.3"
tensorflow = "^2.13.0"
tensorflow-io-gcs-filesystem = "0.37.0"
transformers = "^4.41.1"
scipy = "^1.13.1"
torch = "^2.3.0"
tf-keras = "^2.16.0"

[tool.poetry.dependencies."urllib3"]
version = ">=1.25.4,<2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
